+++ importmap.php
@@ -27,6 +27,8 @@ return [
     '@hotwired/stimulus' => ['version' => '3.2.2'],
     '@symfony/stimulus-bundle' => ['path' => './vendor/symfony/stimulus-bundle/assets/dist/loader.js'],
     '@hotwired/turbo' => ['version' => '8.0.23'],
+    'chart.js/auto' => ['version' => '4.5.1'],
+    '@kurkle/color' => ['version' => '0.3.4'],
     'tom-select' => ['version' => '2.6.1'],
     '@orchidjs/sifter' => ['version' => '1.1.0'],
     '@orchidjs/unicode-variants' => ['version' => '1.1.2'],